pyproject.toml
[tool.poetry]
name = "kingdom"
version = "0.1.0"
description = "Character Sheet for P2E"
authors = ["Telpegaraf <telpegaraf@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
psycopg2 = "^2.9.9"
aiosqlite = "^0.20.0"
pydantic-settings = "^2.3.1"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
alembic = "^1.13.1"
black = "^24.4.2"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
